VAE and comparing 3 methods
initial commit of 2 jupyter notebooks.
One implements a VAE trained on the FashionMNIST dataset, the second compares three different ML pipelines for the wine quality dataset.
In the first, a NN is fit directly on the data, in the second the data is first compressed using PCA and only then fed into a NN, the third approach first trains an Autoencoder and then fits a NN on its compressed representation of the data.